requirements.txt
discord.py==2.4.0
aiohttp==3.11.11
aiohsignal==1.3.2
Flask==3.1.0
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.1
Jinja2==3.1.3
Werkzeug==3.0.1
MarkupSafe==2.1.5
itsdangerous==2.1.2
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
urllib3==2.2.1
colorama==0.4.6
typing-extensions==4.11.0
async-timeout==4.0.3
six==1.16.0
pytz==2024.1
tzdata==2024.1
click==8.1.8
schedule==1.2.1
httpx==0.27.0
loguru==0.7.2
rich==13.7.0
pydantic==2.7.1
tenacity==8.2.3
fastapi==0.110.2
uvicorn==0.29.0
starlette==0.37.2
beautifulsoup4==4.12.3
lxml==5.2.1
yarl==1.9.4
multidict==6.0.5